updated database and some new entry

+++ src/main/java/ca/parimal/connectz/model/entities/EntryKey.java
@@ -10,31 +10,31 @@ import java.io.Serializable;
 public class EntryKey implements Serializable {
 
     @Column(name = "user_id")
-    Integer userId;
+    Long userId;
 
     @Column(name = "media_id")
-    Integer mediaId;
+    Long mediaId;
 
     public EntryKey() {}
 
-    public EntryKey(Integer userId, Integer mediaId) {
+    public EntryKey(Long userId, Long mediaId) {
         this.userId = userId;
         this.mediaId = mediaId;
     }
 
-    public Integer getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(Integer userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
 
-    public Integer getMediaId() {
+    public Long getMediaId() {
         return mediaId;
     }
 
-    public void setMediaId(Integer mediaId) {
+    public void setMediaId(Long mediaId) {
         this.mediaId = mediaId;
     }
